M-sized-integers step 5: a literal takes its width from context, and overflow aborts at all eight — docs/debrief/QUEUE.md
    a: u8 @ 200 · b: i32 @ -70000 · c: u64 @ 42        ->  200 -70000 42
    m: u64 @ 18446744073709551615                      ->  18446744073709551615
    small: u8 @ 300   ->  error[int_out_of_range]: `300` does not fit a `u8`
    b: u8 @ 255 ; b @ b + 1                            ->  panic: integer overflow

Panel 042 said this is where the widths become usable rather than merely
present. Six languages let a literal take its neighbour's type; the one that
does not is Java, and it does it HALF — `byte b = 42;` compiles and `b = b + 1;`
does not. That shape is worse than either whole one here, because the two lines
are one token apart and the diagnostic would have to explain a promotion lattice
§4.3 says does not exist. Full adoption, then.

The arithmetic cost ZERO new lines, as the compiler-engineer predicted:
__builtin_add_overflow judges against the DESTINATION type, so making the local
a uint8_t was the whole change.

Four things the plan did not have, each found by running the thing:

- The decoder needed i128, and not out of caution. `u64` tops at
  18446744073709551615 and `i8` bottoms at -128; no 64-bit integer holds both
  ends. Same for `Const::Int`, which now carries the value from source to
  emitter without wrapping the literal a reader is likeliest to write on
  purpose.
- A NEGATIVE literal is a literal, and the sign has to be folded in before the
  range check: `-128` fits an i8 and `128` does not, so a check seeing them
  separately would refuse the only way to write that type's lowest value. It is
  folded again at lowering, because left as a negation the `128` lands in an
  int8_t before the minus reaches it (clang's -Wconstant-conversion).
- The digits UNDER the minus need the type recorded too, or the lowering finds
  none and the emitter writes `HeroValue` — a clang error naming a type nobody
  wrote.
- print(u64) was still -1. I had added hero_print_uint to the runtime in step 4
  and not wired the emitter's print path to it. Panel 041's finding 1 said
  SIZE_MAX was unreadable under every option on its menu; it reads correctly
  now.

TWO DIAGNOSTICS BECAME ONE, and the second was mine from an hour earlier.
`literal_too_wide` and `int_out_of_range` are the same finding with a different
number in it: a literal with no annotation is measured against `i64`, which is a
width — the default one. A second code would have split one class in two on the
accident of whether the reader had written the type down. `int_out_of_range`
survives, width-aware, and ANNOTATED needs no entry.

AND A REGRESSION I CAUSED AND CAUGHT. Making a literal contextual made `1.5 & 1`
produce two diagnostics — `&` takes `i64`, found `f64`, then `expected f64,
found i64` from the literal being pushed at a type it can never have. Two
diagnostics for one mistake is a defect this compiler has fixed three times, and
adopting the widths reintroduced it within the hour. The rule that fixes it is
`adopts()`: a number literal adopts an integer width and nothing else, while the
other contextual forms adopt anything, because a `.case` or a `???` has no type
at all and a number already has a value.

532 tests green, clippy clean.

+++ crates/heroes/src/ir/inst.rs
@@ -144,7 +144,12 @@ pub enum Abort {
 
 #[derive(Clone, Copy, PartialEq, Debug)]
 pub enum Const {
-    Int(i64),
+    /// **`i128`, and it has to be.** A literal's value must survive from the
+    /// source to the emitter, and after M-sized-integers the widths span
+    /// `i8`'s -128 to `u64`'s 18446744073709551615 — no 64-bit integer holds
+    /// both ends, so a narrower carrier here would silently wrap the one
+    /// literal a reader is most likely to write on purpose (`0xffffffffffffffff`).
+    Int(i128),
     Float(f64),
     Bool(bool),
     Str(StrId),
